pyproject.toml
[project]
name = "nsfw-detector"
version = "0.1.0"
description = ""
authors = [
    {name = "CeliaZHANGPerso",email = "celiazhang349@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard]==0.118.2",
    "uvicorn==0.37.0",
    "pillow==11.3.0",
    "numpy==2.0.2",
    "pandas==2.2.2",
    "tqdm==4.67.1",
    "symspellpy==6.9.0",
    "paddleocr==3.2.0",
    "paddlepaddle==3.2.0",
    "detoxify==0.5.2",
    "torch==2.8.0",
    "transformers==4.57.0",
    "python-multipart==0.0.20",
    "opencv-python==4.12.0.88",
    "importlib==1.0.4",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
